Create EmbedVersion.html
Embeddable styled version 1.0
Needs depth of survey, tested, and funded.
It will then be presented as a paid survey  opportunity to 
people on Mturk as a HIT to get fresh 2025 data about people's preferences about topics associated with charitable contributions.
The focal point for wanting the data is to see if there is a
true problem with the mailers and freebies, or if there is not
enough of a need yet to address that by using the app I am 
developing called Philanthropt-out.  Survey will run under requester 
Milleflori-Data. If that is unavailable, I will generate a new requester name. I will offer .10 per survery. Survey will be one page long and not too invasive.